requirements.txt
google-genai
feedparser
arxiv
jinja2
python-dateutil
certifi
httpx
